pyproject.toml
[project]
name = "litethinking"
version = "0.1.0"
description = ""
authors = [
    {name = "Sebastian Hogar",email = "juan.romero@eduxperia.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "django (>=6.0,<7.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "reportlab (>=4.2.5,<5.0.0)"
]

[tool.poetry]
packages = [{include = "litethinking", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
